Added possibility to resize the window. (min size is 500x300)
When window is resized, we should calculate new size of netting square. Then we should resize view according to size of window. After that we change the view center to display correctly all we need (field, buttons, warning messages).

Defined the size of text labels using the size of netting square, instead of pixels

Used materials:
https://en.sfml-dev.org/forums/index.php?topic=17747.0

https://www.sfml-dev.org/tutorials/2.5/graphics-view.php