Added possibility to save/load game in/from file.
Explanation of process: when you save game, really only sequence of moves that lead to current situation on board are saving. So after reading a game (sequence of moves) we must "play" (or apply) these moves.

For that added methods:
saveMovesHistory(), that writes in file MovesHistory
loadMovesHistory()
applyMoves() - apply all loaded moves and assing variables with correct values according to loaded moves.
checkForWinner() - check and return who is winner (-1, 0, 1)

I used WinAPI functions (GetSaveFileName() and GetOpenFileName) to get path for files. Some sources are in code as a comments.

Moreover, buttons "Save" and "Load" are added.