Made using a buttons more comfortable.
For it added struct Ui::Button that logically describe a button in UI.
Also implemented method isContains() that show if button contains a point. It make easier checking if user clicked in that button.

Added a list of buttons on main window into class Ui. Because of it drawing and checking if some button clicked reduced to iterating through this list and call one method.

Adding a new button or make some changes in existing become much simpler: you need to make changes only in one place - at Ui constructor.

Of course, removed respective code duplications.