Implemented bechmark for matrices multiplication
Why so a big difference between Strassen time at size=1024 and at size=1174? Because according to algo we should expand matrix to nearest degree of 2 (number 2^k).
And when size is 1024 we multiply two matrices size of 1024. But when size is 1174, in fact, we multiply two matrices size of 2048.

Also added some comments for Doxygen